refactor: give mailbox attachment files one owner
